Document Sukkot Stripe endpoints and co-locate endpoint data.
Add docs for create-session, webhook, DB persistence, and Admin health checks; move Stripe donation repository code under Endpoints/Data. Fixes #212.
